+++ wipesrc/wipeout98/file.h
@@ -1,7 +1,7 @@
-long FileLength( char* name );
+int32_t FileLength( char* name );
 
-long LoadFile( char* name, char* buffer );
+int32_t LoadFile( char* name, char* buffer );
 
-void SaveFile( char* name, char* buffer, long length );
+void SaveFile( char* name, char* buffer, int32_t length );
 
 void InitCDFS();
